onboarding-service/requirements.txt
fastapi
uvicorn
sqlalchemy
boto3
pydantic
psycopg2-binary
python-dotenv
botocore[crt]
